move pod scale logic to new file
